+++ src/blocks.js
@@ -146,19 +146,19 @@
 
 /*global Array, BoxMorph,
 Color, ColorPaletteMorph, FrameMorph, Function, HandleMorph, Math, MenuMorph,
-Morph, MorphicPreferences, Object, Point, ScrollFrameMorph, ShadowMorph, ZERO,
-String, StringMorph, TextMorph, contains, degrees, detect, PianoMenuMorph,
-document, getDocumentPositionOf, isNaN, isString, newCanvas, nop, parseFloat,
-radians, useBlurredShadows, SpeechBubbleMorph, modules, StageMorph, Sound,
+Morph, MorphicPreferences, Object, ScrollFrameMorph, ShadowMorph, ZERO, Sound,
+String, StringMorph, TextMorph, contains, degrees, detect, PianoMenuMorph, nop,
+document, getDocumentPositionOf, isNaN, isString, newCanvas, parseFloat, isNil,
+radians, useBlurredShadows, SpeechBubbleMorph, modules, StageMorph, SymbolMorph,
 fontHeight, TableFrameMorph, SpriteMorph, Context, ListWatcherMorph, Rectangle,
 DialogBoxMorph, BlockInputFragmentMorph, PrototypeHatBlockMorph, WHITE, BLACK,
-Costume, IDE_Morph, BlockDialogMorph, BlockEditorMorph, localize, isNil, CLEAR,
+Costume, IDE_Morph, BlockDialogMorph, BlockEditorMorph, localize, CLEAR, Point,
 isSnapObject, PushButtonMorph, SpriteIconMorph, Process, AlignmentMorph,
-CustomCommandBlockMorph, SymbolMorph, ToggleButtonMorph, DialMorph*/
+CustomCommandBlockMorph, ToggleButtonMorph, DialMorph, SnapExtensions*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.blocks = '2021-June-11';
+modules.blocks = '2021-June-14';
 
 var SyntaxElementMorph;
 var BlockMorph;
@@ -351,6 +351,11 @@ SyntaxElementMorph.prototype.labelParts = {
             '(3) sawtooth' : 3,
             '(4) triangle' : 4
         }
+    },
+    '%prim': {
+        type: 'input',
+        tags: 'read-only static',
+        menu: 'primitivesMenu'
     },
     '%audio': {
         type: 'input',
@@ -9319,6 +9324,16 @@ InputSlotMorph.prototype.messagesReceivedMenu = function (searching) {
             null,
             this.world()
         );
+    return dict;
+};
+
+InputSlotMorph.prototype.primitivesMenu = function () {
+    var dict = {},
+        allNames = Array.from(SnapExtensions.keys());
+
+    allNames.sort().forEach(name =>
+        dict[name] = name
+    );
     return dict;
 };